Change int to short
Hello,

i have change every int to short, because teensy use int as 32bit and mega as 16bit.
So the Storage are mismatched/moved to other variable like trigger wheel and something else
http://speeduino.com/forum/viewtopic.php?f=11&t=889

On Mega is short and int the same
https://www.arduino.cc/en/Reference/Short
https://www.arduino.cc/en/Reference/Int

+++ comms.h
@@ -29,8 +29,8 @@ const char pageTitles[] PROGMEM //This is being stored in the avr flash instead 
   };
   
 void command();//This is the heart of the Command Line Interpeter.  All that needed to be done was to make it human readable.
-void sendValues(int packetlength, byte portnum);
-void receiveValue(int offset, byte newValue);
+void sendValues(short packetlength, byte portnum);
+void receiveValue(short offset, byte newValue);
 void saveConfig();
 void sendPage(bool useChar);
 void receiveCalibration(byte tableID);
